pyproject.toml
[project]
name = "notion-fitness-tracker"
version = "0.1.0"
description = "Sync fitness data from Hevy, Strava, and Garmin to Notion"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "garminconnect>=0.2.25",
    "python-dotenv>=1.0.0",
]

[project.scripts]
hevy-sync = "scripts.hevy_sync:main"
garmin-sync = "scripts.garmin_sync:main"
dashboard-update = "scripts.update_dashboard:main"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "types-requests>=2.31.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "ANN", "RET"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
